complete overhaul of map file location
